test: adjust to safe commit append
When we create a commit and tell the library to update a ref, we need to
make sure to pass the current tip.

This is what we should have been doing in a function called
'upateReadme()' anyway. The existing code creates a new root commit,
which is not an update.