Link dynamically or statically depending on build flags
This lets us specify which set of rules to link to libgit2 we'll use during
build. Together with the changes to the Makefile, this lets you specify

    make install

or

    make install-static

to build either dynamically against the system-provided library or against the
version specified in the submodule.

The dynamic linking against the system library is still the default in order to
allow for the Go toolchain to build it when it's used as a dependency.