rebase: correct the return values for CurrentOperationIndex
We were incorectly reporting `C.GIT_REBASE_NO_OPERATION` as an error code when
it is none. We should instead return it as the value. The compiler doesn't seem
to actually look at the sizes so instead we must recreate the value ourselves
with `^uint(0)`.

The error return is kept for API compatibility but should go away eventually.